Don't set -fobject-code for GHCI, fix variable binding
This _was_ enabled as it can help when loading a module via
cabal repl and expecting file-watching to work in GHCID.

As our project is very simple we can just `-isrc` and `-itest` which
appears to be enough to get file-watching.

Nice side effect is it's now _extremely_ fast to recompile as there's
no writing to disk. :D