Increase base upper bound to work with newer platform installs
This should allow Ben to run this on windows with only the haskell
platform (at recent versions).

I still recommend the MacOS /Linux + Nix env setup ... only because
that's the one I use, and now most of us are using.

If not using Nix I'd recommend using Cabal's sandbox feature as outlined
here:
https://www.haskell.org/cabal/users-guide/installing-packages.html#developing-with-sandboxes